Got auth to work and began integrating tokens into it
