+++ python/neuroglancer/default_credentials_manager.py
@@ -12,9 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from . import credentials_provider, google_credentials
+from . import credentials_provider, google_credentials, boss_credentials
 
 default_credentials_manager = credentials_provider.CredentialsManager()
+boss_credentials_provider = boss_credentials.BossCredentialsProvider()
 default_credentials_manager.register(
     u'google-brainmaps',
     lambda _parameters: google_credentials.GoogleOAuth2FlowCredentialsProvider(
@@ -29,5 +30,8 @@ default_credentials_manager.register(
 
 default_credentials_manager.register(
     u'boss',
-    lambda _parameters: google_credentials.BossCredentialsProvider()
+    lambda _parameters: boss_credentials_provider
 )
+
+def set_boss_token(token):
+    boss_credentials_provider.set_token(token)
