+++ src/neuroglancer/datasource/boss/api.ts
@@ -40,7 +40,13 @@ export function fetchWithBossCredentials<T>(
       credentialsProvider, input, init, transformResponse,
       credentials => {
         const headers = new Headers(init.headers);
-        headers.set('Authorization', `Bearer ${credentials}`);
+        if (typeof(credentials) === "string") {
+          headers.set('Authorization', `Bearer ${credentials}`);
+        }
+        else {
+          //@ts-ignore
+          headers.set('Authorization', `${credentials.tokenType} ${credentials.accessToken}`); 
+        }
         return {...init, headers};
       },
       error => {
